refactor(services): move httpClient to core directory
